test(openai): move OpenAI provider tests to dedicated file
